requirements.txt
pandas
numpy
pydantic
python-dotenv
pytest
requests
python-ulid
websockets